requirements.txt
absl-py==1.4.0
altair==4.2.2
astunparse==1.6.3
attrs==22.2.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
flatbuffers==23.3.3
fonttools==4.39.4
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.3.0
jax==0.4.8
Jinja2==3.1.2
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
ml-dtypes==0.1.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==1.5.3
Pillow==9.5.0
plotly==5.14.1
protobuf==3.20.3
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.8.0
Pygments==2.15.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.4
rsa==4.9
scipy==1.10.1
six==1.16.0
smmap==5.0.0
streamlit==1.21.0
streamlit-cropper==0.2.0
tenacity==8.2.2
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.15.0
